+++ src/preload/index.d.ts
@@ -5,7 +5,16 @@ interface IElectronAPI {
   shortcuts: {
     onSettings: (callback: () => void) => () => void
     onNewApplication: (callback: () => void) => () => void
-    onBuildFromScratch: (callback: () => void) => () => void
+    onSave: (callback: () => void) => () => void
+    onExportPdf: (callback: () => void) => () => void
+    onDocumentSettings: (callback: () => void) => () => void
+    onPrevApp: (callback: () => void) => () => void
+    onNextApp: (callback: () => void) => () => void
+    onLatestApp: (callback: () => void) => () => void
+    onOldestApp: (callback: () => void) => () => void
+    onTailor: (callback: () => void) => () => void
+    onToggleSidebar: (callback: () => void) => () => void
+    onToggleColumns: (callback: () => void) => () => void
   }
   platform: NodeJS.Platform
   settings: {